Controller Payment

+++ backend/Model/Dtos/PaymentDtos/PaymentCreateDto.cs
@@ -3,8 +3,8 @@
     public class PaymentCreateDto
     {
         public int Reservation_id { get; set; }
-        public int MiscCharges { get; set; }
-        public int Total { get; set; }
+        public decimal MiscCharges { get; set; }
+        public decimal Total { get; set; }
         public string PaymentMethod { get; set; }
     }
 }
