+++ backend/Model/Entity/Payment.cs
@@ -5,8 +5,8 @@
         public int Id { get; set; }
         public int Reservation_id { get; set; }
         public Reservation Reservation { get; set; }
-        public int MiscCharges { get; set; }
-        public int Total { get; set; }
+        public decimal MiscCharges { get; set; }
+        public decimal Total { get; set; }
         public string PaymentMethod { get; set; }
     }
 }
